pyproject.toml
[project]
name = "botty-framework"
version = "0.0.1"
description = "Botty brings the elegance of FastAPI's dependency injection and type hints to Telegram bot development. Write clean, testable bot code with automatic dependency resolution, built-in message registry, and a developer-friendly API."
readme = "README.md"
authors = [
    { name = "melaveetha", email = "melaveetha@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "python-telegram-bot>=22.6",
    "sqlmodel>=0.0.32",
]

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "botty"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-loguru>=0.4.0",
    "ruff>=0.15.0",
    "ty>=0.0.15",
]

[tool.uv.workspace]
members = [
    "examples/task-manager-bot",
]

[tool.ty.src]
exclude = [
    "examples/**"
]

[tool.ruff.lint]
ignore=["E402"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
